Refactors code on classes
