+++ opensrp-giz-malawi/src/main/java/org/smartregister/giz/configuration/GizOpdRegisterRowOptions.java
@@ -39,7 +39,6 @@ public class GizOpdRegisterRowOptions implements OpdRegisterRowOptions {
         } else {
             dueButton.setBackgroundResource(R.color.transparent);
             dueButton.setText(" ");
-
         }
     }
